# Conflicts:
#	Modules/Utilities/Resources/views/page/table-sub/slider.blade.php

+++ Modules/Utilities/Http/Controllers/UploadController.php
@@ -18,6 +18,8 @@ class UploadController extends Controller
 
     public $uploadDirectory = '';
 
+    public $stopRelationSave = false;
+
     public $relationType = '';
 
     function __construct()
@@ -32,10 +34,25 @@ class UploadController extends Controller
             ? $this->imageLocalConfig['relationType']
             : $this->imageGeneralConfig['relationType'];
 
-        // get path directory storage
-        $this->uploadDirectory = isset($this->imageLocalConfig['upload_directory'])
+        // get path upload directory storage
+        $mainDirectory = isset($this->imageGeneralConfig[$routeParam['type']]['main_directory'])
+            ? $this->imageGeneralConfig[$routeParam['type']]['main_directory']
+            : 'app';
+
+        // get path upload directory storage
+        $uploadDirectory = isset($this->imageLocalConfig['upload_directory'])
             ? $this->imageLocalConfig['upload_directory']
             : $this->imageGeneralConfig[$routeParam['type']]['upload_directory'];
+
+        // stop all relation oper
+        $this->stopRelationSave = isset($this->imageLocalConfig['stopRelationSave'])
+            ? $this->imageLocalConfig['stopRelationSave']
+            : false;
+
+        $folderUpload = Str::plural($routeParam['model']);
+
+        $this->targetDirectory = "$uploadDirectory\\$folderUpload";
+        $this->uploadDirectory = "$mainDirectory\\$uploadDirectory\\$folderUpload";
     }
 
     function index(Request $request ,$model ,$type) {
@@ -43,17 +60,20 @@ class UploadController extends Controller
         return 'done';
     }
 
-    function upload(UploadFormRequest $request ,$model ,$type){
+    function upload(UploadFormRequest $request ,$model ,$type) {
 
         $file = $request->file($model);
         $file = is_array($file) ? $file[0] : $file;
 
-        // to do : from this we will know what type of image is original or cropped or thumpnail
+        // todo : from this we will know what type of image is original or cropped or thumpnail
         // imageType this must stored in db
         $imageType = 'original';
 
         $path       = storage_path($this->uploadDirectory);
         $hashName   = strtolower(str_random(12))."_{$imageType}_". $file->getClientOriginalName();
+
+        // make directory if not exists
+        Storage::makeDirectory($this->targetDirectory);
 
         // move with intervention
         $imgRezise = \Image::make($file->getRealPath());
@@ -75,41 +95,49 @@ class UploadController extends Controller
         $partFunc = Str::studly($type);
         $returnParam = $this->{"saveUpload{$partFunc}Db"}($extraParams);
 
-        // get config model
-        $dbModel = $this->imageLocalConfig['model'];
+        if(!$this->stopRelationSave)
+        {
+            // get config model
+            $dbModel = $this->imageLocalConfig['model'];
 
-        // save relation file
-        $dbModel = $dbModel::findOrFail($request->get('id'));
-        // relationType has to be many or one
-        if($this->relationType == 'many')
-            $dbModel->image()->attach($returnParam->id);
-        else
-            $dbModel->update(["{$type}_id" => $returnParam->id]);
+            // save relation file
+            $dbModel = $dbModel::findOrFail($request->get('id'));
+            // relationType has to be many or one
+            if($this->relationType == 'many')
+                $dbModel->image()->attach($returnParam->id);
+            else
+                $dbModel->update(["{$type}_id" => $returnParam->id]);
+        }
 
         return array_merge(["success" => true], [$model => $returnParam]);
     }
 
     function destroy(Request $request ,$model ,$type) {
 
-        // get config model
-        $dbModel = $this->imageLocalConfig['model'];
+        request()->request->add(['transSaveOper' => false]);
 
-        // save relation file
-        $dbModel = $dbModel::findOrFail($request->get('id'));
+        if(!$this->stopRelationSave)
+        {
+            // get config model
+            $dbModel = $this->imageLocalConfig['model'];
 
-        // relationType has to be many or one
-        if($this->relationType == 'many')
-            $dbModel->image()->detach($request->get('key'));
-        else
-            $dbModel->update(["{$type}_id" => null]);
+            // save relation file
+            $dbModel = $dbModel::findOrFail($request->get('id'));
+
+            // relationType has to be many or one
+            if($this->relationType == 'many')
+                $dbModel->image()->detach($request->get('key'));
+            else
+                $dbModel->update(["{$type}_id" => null]);
+        }
 
         //delete file from db
         $partFunc = Str::studly($type);
         $this->{"destroyUpload{$partFunc}Db"}($request->get('key'));
 
-        // set
+        // set storage path for file delete
         $path = storage_path($this->uploadDirectory."\\".$request->get('file_name'));
-        // delete image from storage .. ps: this accept just file name bust i pass full path.
+        // delete image from storage .. ps: this accept just file name but i pass full path.
         Storage::delete($path);
         // remove symbolic link from image
         if(is_file($path))
